Add ICASSP25 audio generation and fix MPS compatibility issues
Creates comprehensive audio generation pipeline for ICASSP25 with FTM synthesis:

Audio Generation:
- Add icassp25/01_generate_audio.py: New script for FTM modal drum synthesis
- Update Makefile ai25 target: Generate data instead of symlinking to ICASSP23
- Generates 100k samples matching ICASSP25 parameter dataset structure

MPS Compatibility Fixes:
- Fix src/pnp_synth/physical/ftm.py: Change float64 to float32 in pi tensor creation
- Fix src/pnp_synth/neural/cnn.py: Replace .double() with .float() in training step
- Resolves "Cannot convert MPS Tensor to float64" errors on Apple Silicon

The generation script successfully creates HDF5 files with proper structure for
train/val/test folds, enabling make ri25pl to progress through model initialization
and into actual training phases. Addresses the core issue where ICASSP25 had
100k parameter dataset but no corresponding audio data.

+++ src/pnp_synth/neural/cnn.py
@@ -154,10 +154,10 @@ class EffNet(pl.LightningModule):
 
     def step(self, batch, fold, batch_idx):
         Sy = batch['feature'].to(self.current_device)
-        y = batch['y'].to(self.current_device).double()
+        y = batch['y'].to(self.current_device).float()
         weight = batch['weight'].to(self.current_device)
         try:
-            M = batch['M'].to(self.current_device).double()
+            M = batch['M'].to(self.current_device).float()
             M_mean = batch['M_mean'].to(self.current_device)
             self.LMA_lambda0 = batch['lambda0'].to(self.current_device)
             self.LMA_threshold = self.LMA_lambda0 # set threshold to be the initialized lambda
